% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/create.sdf.fast.R
\name{create.sdf.fast}
\alias{create.sdf.fast}
\title{Transfrom an array of data into a gridded spatial data frame}
\usage{
create.sdf.fast(lat.prod, lon.prod, prod, box.size = 2, name.prod, agg.fun)
}
\arguments{
\item{lat.prod}{Array; Latitudes of the dataset.}

\item{lon.prod}{Array; Longitudes of the dataset.}

\item{prod}{Array; Dataset to be represented in the spatial dataframe.}

\item{box.size}{Numeric; Size of the gridbox where the data is averaged. Default is set to 2.0 degrees.}

\item{name.prod}{Character; Name of the data}
}
\description{
This function transforms an array of latlon data into a gridded latlon data frame. The data is averaged in a gridbox.
}
\examples{
To come.
}
\keyword{grid,}
\keyword{map}

