% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/grid.list.R
\name{grid.list}
\alias{grid.list}
\title{Grid an array of latlon dataset.}
\usage{
grid.list(lon, lat, z, box.size = 2)
}
\arguments{
\item{lon}{Array; Longitudes of the dataset.}

\item{lat}{Array; Latitudes of the dataset.}

\item{z}{Array; Dataset to be represented in the spatial dataframe.}

\item{box.size}{Numeric; Size of the gridbox where the data is averaged. Default is set to 2.0 degrees.}
}
\description{
This function returns a list of gridded dataset.
}
\examples{
To come.
}
\keyword{grid}

