% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/plot.proj.map.R
\name{plot.proj.map}
\alias{plot.proj.map}
\title{Plot a projected map of a given dataset}
\usage{
plot.proj.map(Lat, Lon, Prod, proj = "eck4", sampling = "cubic",
  interp = TRUE, overwrite = TRUE, dir.GTiff = ".", color = "",
  ncol = 64, box.size = 2, ...)
}
\arguments{
\item{Lat}{Array; Latitudes of the dataset.}

\item{Lon}{Array; Longitudes of the dataset.}

\item{Prod}{Array; Dataset to be represented in the spatial dataframe.}

\item{proj}{Character; Name of the output projection (correspond to the names of rgdal PROJ4 projections, call projInfo for more information). Current supported projections are: aeqd, cass, eck1-2-3-4-5-6, eqc, gall, goode, igh, longlat, mill, moll, poly, robin, siny, vandg, wag1-2-3-4-5-6. Default is set to eck4.}

\item{sampling}{Character; Type of resampling (near, bilinear, cubic, cubicspline, lanczos, average, mode). Default is set to cubic.}

\item{interp}{Logical; Activate the interpolation of the dataset during resampling. Default is TRUE.}

\item{overwrite}{Logical; If FALSE, the function tries to re-use an existing GTiff file, otherwise a new one is created. Default is TRUE.}

\item{color}{Array/Character; The input can be a color palette or the name of a ColorBrewer palette. Default is the Spectral type of ColorBrewer.}

\item{ncol}{Numeric; Number of colors in the output colorbors. The input palette is interpolated.}

\item{box.size}{Numeric; Size of the gridbox where the data is averaged. Default is set to 2.0 degrees.}

\item{...}{Additional arguments passed to image.plot (fields package).}

\item{dir.Gtiff}{Character; Directory where is stored the raster dataset in a GTiff format. Default is ".".}
}
\description{
This function returns plots of a mapped given dataset. The initial dataset is averagged into a grid before being resampled, interpolated and projected.
}
\examples{
To come.
}
\keyword{map}
\keyword{projection,}

