% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/project.map.R
\name{project.map}
\alias{project.map}
\title{Project a spatial data frame}
\usage{
project.map(proj = "eck4")
}
\arguments{
\item{proj}{Character; Name of the output projection (correspond to the names of rgdal PROJ4 projections, call projInfo for more information). Current supported projections are: aeqd, cass, eck1-2-3-4-5-6, eqc, gall, goode, igh, longlat, mill, moll, poly, robin, siny, vandg, wag1-2-3-4-5-6. Default is set to eck4.}
}
\description{
This function returns a projected map
}
\examples{
To come.
}
\keyword{map}
\keyword{projection,}

