% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/project.sdf.R
\name{project.sdf}
\alias{project.sdf}
\title{Project a spatial data frame}
\usage{
project.sdf(sdf.latlon, proj = "eck4", sampling = "cubic",
  dir.GTiff = ".", interp = TRUE, overwrite = TRUE, name.prod = "")
}
\arguments{
\item{proj}{Character; Name of the output projection (correspond to the names of rgdal PROJ4 projections, call projInfo for more information). Current supported projections are: aeqd, cass, eck1-2-3-4-5-6, eqc, gall, goode, igh, longlat, mill, moll, poly, robin, siny, vandg, wag1-2-3-4-5-6. Default is set to eck4.}

\item{sampling}{Character; Type of resampling (near, bilinear, cubic, cubicspline, lanczos, average, mode). Default is set to cubic.}

\item{interp}{Logical; Activate the interpolation of the dataset during resampling. Default is TRUE.}

\item{overwrite}{Logical; If FALSE, the function tries to re-use an existing GTiff file, otherwise a new one is created. Default is TRUE.}

\item{name.prod}{Character; Name of the data}

\item{sds.latlon}{Spatial gridded dataframe; Input data in longlat projection.}

\item{dir.Gtiff}{Character; Directory where is stored the raster dataset in a GTiff format. Default is ".".}
}
\description{
This function projects a longlat dataframe into a given projection. The data can also be resampled and interpolated.
}
\examples{
To come.
}
\keyword{projection}

